Fix nullification of message const& in socket::send()
zmq_msg_send() nullifies the message that is passed in. The function had
a signature taking a message const&, which should not be modified.
To fix this, the signature is changed to pass the message by value, so a
copy is made if necessary.

This commit also removes some unnecessary const_casts. One of these
would have correctly complained about passing a zmq_msg_t const* to
zmq_msg_send().
The const_cast for the src-argument of zmq_msg_copy() as well as various
accessors is necessary though; libzmq doesn't seem to be stringent on
type-based const-correctness.